pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eml-research-poc"
version = "0.1.0"
description = "Research POC: train/evaluate an AI agent that reasons about math problems using the EML (Exp-Minus-Log) primitive as a tool."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "ElMatiOfficial" }]
dependencies = [
    "anthropic>=0.40.0",
    "sympy>=1.12",
    "eml-translator @ git+https://github.com/ElMatiOfficial/EML-Matemathical-Translator-for-AI.git@main",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
eml-benchmark = "eml_research.benchmark:main"
eml-tool = "eml_research.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
